+++ crates/kct_package/src/compiler/subpackage.rs
@@ -1,4 +1,5 @@
-use super::{Release, INCLUDE_PARAM, SUBPACKAGES_FOLDER};
+use super::Compiler;
+use super::{Release, INCLUDE_PARAM};
 use crate::Package;
 use jrsonnet_evaluator::{
 	error::Error as JrError, error::LocError, native::NativeCallback, FuncVal, Val,
@@ -7,13 +8,13 @@ use jrsonnet_parser::{Param, ParamsDesc};
 use serde_json::Value;
 use std::{convert::TryFrom, rc::Rc};
 
-pub fn create_function(pkg: &Package, release: &Option<Release>) -> Val {
+pub fn create_function(compiler: &Compiler, release: &Option<Release>) -> Val {
 	let params = ParamsDesc(Rc::new(vec![
 		Param("name".into(), None),
 		Param("input".into(), None),
 	]));
 
-	let root = pkg.root.clone().join(SUBPACKAGES_FOLDER);
+	let vendor = compiler.vendor.clone();
 	let release = release.clone();
 	let render = move |_caller, params: &[Val]| -> std::result::Result<Val, LocError> {
 		let name = params.get(0).unwrap();
@@ -25,8 +26,9 @@ pub fn create_function(pkg: &Package, release: &Option<Release>) -> Val {
 				)))
 			}
 		};
-		let mut root = root.join(&package.to_string());
-		root.set_extension("tgz");
+		let root = vendor.join(&package.to_string());
+		let mut subcompiler = Compiler::new(&root);
+		subcompiler.vendor = vendor.clone();
 
 		let package = Package::try_from(root)
 			.map_err(|err| LocError::new(JrError::RuntimeError(err.to_string().into())))?;
@@ -36,8 +38,12 @@ pub fn create_function(pkg: &Package, release: &Option<Release>) -> Val {
 			.map(|val| val.to_string().unwrap())
 			.map(|val| serde_json::from_str(&val).unwrap());
 
-		let rendered = package
-			.compile(input, release.clone())
+		package
+			.validate_input(&input)
+			.map_err(|err| LocError::new(JrError::RuntimeError(err.to_string().into())))?;
+
+		let rendered = subcompiler
+			.compile(package, input.unwrap_or(Value::Null), release.clone())
 			.map_err(|err| LocError::new(JrError::RuntimeError(err.to_string().into())))?;
 
 		Ok(Val::from(&rendered))