+++ CHANGELOG.md
@@ -10,9 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - allows multiple occurrences of `input`
+- enable subpackages to render their own subpackages
 
 ### Changed
 
+- subpackages as normal jsonnet dependencies
 - change from `_.values` to `_.input` on injected global
 - expect `schema.json` instead of `values.schema.json`
 - read `default.json` instead of `values.json` for defaults